Production setting management, testing deployment (#107)
* Add docker-compose setup, keep secrets in run/scionlab.env

Docker compose file starts postgres, redis (for huey), huey, django and caddy as a proxy.
Use volumes to store the persistent data for all the services -- this is supposedly superior to mounts for various operational reasons.
Secrets/site configuration are stored in the unversioned file run/scionlab.env 
```
POSTGRES_DB=...
POSTGRES_USER=...
POSTGRES_PASSWORD=...
# User/password for mail.ethz.ch:
EMAIL_HOST_USER=...
EMAIL_HOST_PASSWORD=...
RECAPTCHA_PUBLIC_KEY=...
RECAPTCHA_PRIVATE_KEY=...
VPN_CA_KEY_PASSWORD=...
SCIONLAB_SITE=https://testing.scionlab.org
```
Additionally the following files are stored/read from run/:
 - SECRET.key: djangos secret, created automatically if none set
 - ssh_config, (and related id_rsa, id_rsa.pub): ssh info for deployment to managed hosts
 - root_ca_cert.pem, root_ca_key.pem: root CA for VPNs

* Add (basic) Caddyfile and configure django behind caddy.

* Add requirements for PostgreSQL, gunicorn, redis

* Install graphviz in anticipation of #112

* Add ribbon to indicate testing mode

* Minor fixes for VPN config installation, add empty ccd/-dir

Simplifies handling in scionlab-config script.
Also set mtime in tar because it looks nicer. Fails test that expects identical tar between requests (disabled check).

* Dont use small integer fields for ports

We use port numbers exceeding the range for *signed* 16-bit integers (which is what smallint is in postgres).

* Use specific ssh_config file for deployment

Allow hostname instead of only ip for SSH hostname

* Fix various issues in scionlab-config

sudo call and dhparam gen 
Handle SSH_ORIGINAL_COMMAND in scionlab-config to allow to restrict
access to a key to running this script in .ssh/authorized_keys:

  command=scionlab-config <...key...> auto-deploy@scionlab.org

* Cleanup host update version bump

* Fix pingpong supervisor conf

* Split integration tests into separate circleci job

Add comments about possible future cleanup.
Also, unify some of the various clones performed.

* Remove leftover file

The values in 'x509_config.default' have been integrated into the normal
django settings some time ago.

+++ scionlab/models/user_as.py
@@ -163,7 +163,7 @@ class UserAS(AS):
     # if she switches to VPN and back.
     public_ip = models.GenericIPAddressField(null=True, blank=True)
     bind_ip = models.GenericIPAddressField(null=True, blank=True)
-    bind_port = models.PositiveSmallIntegerField(null=True, blank=True)
+    bind_port = models.PositiveIntegerField(null=True, blank=True)
 
     installation_type = models.CharField(
         choices=INSTALLATION_TYPES,
